requirements.txt
requests
GitPython
platformio
